+++ server/model/model.go
@@ -3,16 +3,7 @@ package model
 import (
 	"github.com/jinzhu/gorm"
 	_ "github.com/jinzhu/gorm/dialects/mysql"
-
-	"log"
 )
-
-type Post struct {
-	gorm.Model
-	Title string `gorm:"size:255;not null"`
-	Author string `gorm:"size:255;not null"`
-	Contents *string
-}
 
 var database *gorm.DB
 
@@ -23,12 +14,5 @@ func InitDB() {
 		panic("db connection error");
 	}
 	database = db
-
-	var post []Post
-	if err := database.Find(&post).Error; err != nil {
-		log.Fatal(err)
-	}
-
-	log.Print(post)
 
 }
